fixed some styles and moved transaction modal from global state to transactions
