+++ src/redux/transactions/transactions-actions.js
@@ -1,0 +1,3 @@
+import { createAction } from '@reduxjs/toolkit';
+
+export const setIsModalAddTransactionOpen = createAction('transactions/setIsModalAddTransactionOpen')
