+++ src/redux/transactions/transactions-selectors.js
@@ -1,4 +1,6 @@
 export const getTransactions = state => state.transactions.items;
 export const getIsLoading = state => state.transactions.isLoading;
+export const getIsModalAddTransactionOpen = state => state.transactions.isModalAddTransactionOpen;
+
